Update Entire Code.txt
This is the entire code for the project. I did not upload the images, so you will have to make your own. Twitter data by default is turned, so to turn it on scroll down, and there will be a comment; just remove the quotation marks off the string and twitter data can be accessed.